ddd_llm_app_tag6/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ddd_llm_app"
version = "0.1.0"
description = "A tool for Domain-Driven Design analysis and OpenAPI generation."
authors = [
    { name = "Urso Wieske", email = "uwieske@gmail.com" }
]
requires-python = ">=3.12"
readme = "README.md"
license = {"file" = "LICENSE"}
dependencies = [
"anthropic==0.42.0",
"build==1.2.2.post1",
"langchain==0.3.14",
"langchain-anthropic==0.3.1",
"langchain-community==0.3.14",
"langchain-openai==0.3.0",
"numpy==1.26.4",
"omegaconf==2.3.0",
"openai==1.59.6",
"python-dotenv==1.0.1",
"PyYAML==6.0.2",
"sentencepiece==0.2.0",
"torch==2.2.2",
"transformers==4.48.0",
"fastapi",
"uvicorn",
"flask",
"langserve[all]",
"streamlit",
 "faiss-cpu==1.10.0",
 "tiktoken==0.8.0",
]
[project.optional-dependencies]
dev = ["pytest", "flake8", "build==1.2.2.post1", "setuptools==75.7.0"]

[project.urls]
Homepage = "https://github.com/uwieske/ddd_llm_app"
Bug-Tracker = "https://github.com/uwieske/ddd_llm_app/issues"

[tool.setuptools]
packages = { find = { where = ["."], exclude = ["tests*", "tests/*"]  } }
package-dir = { "" = "." }
include-package-data = true

[tool.setuptools.package-data]
ddd_llm_app = [
    "config/default-config.yaml",
    "templates/*.html",
    "static/*"
]

#[tool.setuptools.data-files]
## Ensure templates directory is explicitly included
#data-files = [
#    "templates", "templates/*"
#]
[project.scripts]
ddd_llm_app = "ddd_llm_app.main:main"
ddd_llm_web = "ddd_llm_app.app:main"
ddd_llm_rag_api = "ddd_llm_app.rag_api:start"
#
#[tool.my_setup_hooks]
## Add custom hook configuration for build-time tasks
#post-install = "ddd_llm_app.hooks.post_install"






#[build-system]
#requires = ["setuptools>=61.0", "wheel", "setuptools_scm"]
#build-backend = "setuptools.build_meta"
#
#[project]
#name = "ddd_llm_app"
#version = "0.1.0"
#description = "A tool for Domain-Driven Design analysis and OpenAPI generation."
#authors = [
#    { name = "Urso Wieske", email = "uwieske@gmail.com" }
#]
#requires-python = ">=3.12"
#readme = "README.md"
#license = {"file" = "LICENSE"}
#dependencies = [
#    "langchain",
#    "langchain-community",
#    "openai",
#    "langchain-openai",
#    "anthropic",
#    "langchain-anthropic",
#    "torch",
#    "sentencepiece",
#    "transformers",
#    "numpy<2",
#    "pyyaml",
#    "omegaconf",
#    "python-dotenv",
#    "argparse",
#]
#[project.optional-dependencies]
#dev = ["pytest", "flake8", "build"]
#
#[project.urls]
#Homepage = "https://github.com/uwieske/ddd_llm_app"
#Bug-Tracker = "https://github.com/uwieske/ddd_llm_app/issues"
#
#[tool.setuptools]
#packages = { find = { where = ["."], exclude = ["tests*", "tests/*"]  } }
#package-dir = { "" = "." }
#include-package-data = true
#
#[tool.setuptools.package-data]
#"ddd_llm_app" = [
#    "default-config.yaml",
#    "templates/ddd_prompts.txt"
#]
#[project.scripts]
#ddd_llm_app = "ddd_llm_app.main:main"
#
##
##[tool.my_setup_hooks]
### Add custom hook configuration for build-time tasks
##post-install = "ddd_llm_app.hooks.post_install"
#
